+++ src/html/footer.rs
@@ -1,8 +1,6 @@
 use std::io::Write;
 
-use anyhow::Result;
-
-use crate::html::IntoHtmlNode;
+use crate::html::{IntoHtmlNode, IntoHtmlNodeResult};
 
 #[derive(Debug)]
 pub struct Footer {
@@ -22,7 +20,7 @@ impl Footer {
 }
 
 impl IntoHtmlNode for Footer {
-    fn transform_into_raw_html(&self, buffer: &mut dyn Write) -> Result<()> {
+    fn transform_into_raw_html(&self, buffer: &mut dyn Write) -> IntoHtmlNodeResult<()> {
         writeln!(buffer, "<footer>")?;
 
         for element in &self.elements {
@@ -34,7 +32,7 @@ impl IntoHtmlNode for Footer {
         Ok(())
     }
 
-    fn transform_into_raw_css(&self, buffer: &mut dyn Write) -> Result<()> {
+    fn transform_into_raw_css(&self, buffer: &mut dyn Write) -> IntoHtmlNodeResult<()> {
         for element in &self.elements {
             element.transform_into_raw_css(buffer)?;
         }
@@ -42,7 +40,7 @@ impl IntoHtmlNode for Footer {
         Ok(())
     }
 
-    fn transform_into_raw_js(&self, buffer: &mut dyn Write) -> Result<()> {
+    fn transform_into_raw_js(&self, buffer: &mut dyn Write) -> IntoHtmlNodeResult<()> {
         for element in &self.elements {
             element.transform_into_raw_js(buffer)?;
         }