+++ src/html/meta.rs
@@ -1,9 +1,7 @@
 use std::collections::HashMap;
 use std::io::Write;
 
-use anyhow::Result;
-
-use crate::html::IntoHtmlNode;
+use crate::html::{IntoHtmlNode, IntoHtmlNodeResult};
 
 #[derive(Debug)]
 pub struct Meta {
@@ -23,7 +21,7 @@ impl Meta {
 }
 
 impl IntoHtmlNode for Meta {
-    fn transform_into_raw_html(&self, buffer: &mut dyn Write) -> Result<()> {
+    fn transform_into_raw_html(&self, buffer: &mut dyn Write) -> IntoHtmlNodeResult<()> {
         write!(buffer, "<meta")?;
 
         for (key, value) in self.values.iter() {
@@ -35,11 +33,11 @@ impl IntoHtmlNode for Meta {
         Ok(())
     }
 
-    fn transform_into_raw_css(&self, _buffer: &mut dyn Write) -> Result<()> {
+    fn transform_into_raw_css(&self, _buffer: &mut dyn Write) -> IntoHtmlNodeResult<()> {
         Ok(())
     }
 
-    fn transform_into_raw_js(&self, _buffer: &mut dyn Write) -> Result<()> {
+    fn transform_into_raw_js(&self, _buffer: &mut dyn Write) -> IntoHtmlNodeResult<()> {
         Ok(())
     }
 }